+++ release_notes.md
@@ -1,3 +1,17 @@
+## Release Note 2.37.0
+
+### features
+- Add code in preparation for retirement of old training and ephys local server pipelines
+
+### bugfixes
+- Change handling of trials extraction of iblrigv8 sessions such that pregenerated session is not used
+
+## Release Note 2.36.0
+
+### features
+- Adding spike sorting iblsort task
+
+
 ## Release Note 2.35.0
 
 ### features
@@ -14,6 +28,9 @@
 - Flag to allow session registration without water administrations
 - Support extraction of repNum for advancedChoiceWorld
 - Support matplotlib v3.9; min slidingRP version now 1.1.1
+
+#### 2.35.3
+- Use correct task for timeline acquisitions in make_pipeline
 
 ## Release Note 2.34.0